requirements.txt
# Core requirements
scrapy==2.8.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
hashids==1.3.1
beautifulsoup4==4.11.2
requests==2.28.2

# Dependencies
attrs>=21.3.0
Automat>=0.8.0
certifi>=2023.7.22
cffi>=1.15.1
charset-normalizer>=3.1.0
constantly>=15.1.0
cryptography>=41.0.2
cssselect>=1.2.0
filelock>=3.12.2
hyperlink>=21.0.0
idna>=3.4
incremental>=22.10.0
itemadapter>=0.8.0
itemloaders>=1.1.0
jmespath>=1.0.1
lxml>=4.9.3
packaging>=23.1
parsel>=1.8.1
Protego>=0.2.1
pyasn1>=0.5.0
pyasn1-modules>=0.3.0
pycparser>=2.21
PyDispatcher>=2.0.7
pyOpenSSL>=23.2.0
queuelib>=1.6.2
service-identity>=23.1.0
six>=1.16.0
soupsieve>=2.4.1
Twisted>=22.10.0
typing_extensions>=4.7.1
urllib3>=1.26.16
w3lib>=2.1.1
zope.interface>=6.0
